Actually fixed how saving works. Note: not settled on whether I want boosts to be saved on task finish but I cant figure out how on earth its meant to work.

+++ app/src/main/java/com/mdelsordo/stepquest/model/Boost.java
@@ -26,7 +26,7 @@ public class Boost {
     private String mDesc;
     private String mName;
 
-    public Boost(double multiplier, int duration){
+    public Boost(double multiplier, long duration){
         mDuration = duration;
         mStepMultiplier = multiplier;
         mDesc = makeDesc(mDuration, mStepMultiplier);
